+++ Changelog.md
@@ -1,5 +1,43 @@
 # Mammoth.Extensions.DependencyInjection
 
+## 0.2.0
+
+- AssemblyInspector: added Configure() method [#2](https://github.com/PrimordialCode/Mammoth.Extensions.DependencyInjection/issues/2).
+
+## Breaking Changes
+
+- Namespace changed for the following classes:
+
+  - `Dependency` -> from `Mammoth.Extensions.DependencyInjection` to `Mammoth.Extensions.DependencyInjection.Configuration`
+  - `Parameter` -> from `Mammoth.Extensions.DependencyInjection` to `Mammoth.Extensions.DependencyInjection.Configuration`
+
+- AssemblyInspector lifestyle selectors signature chaged; it does not accept "dependsOn" anymore, use the new Configure() method instead:
+
+  ```csharp
+  var descriptors = new AssemblyInspector()
+  	.FromAssemblyContaining<TestService>()
+  	.BasedOn(typeof(TestService))
+  	.WithServiceBase()
+  	.LifestyleTransient(dependsOn: new Dependency[]
+  	{
+  		Parameter.ForKey("param").Eq("nonexisting")
+  	});
+  ```
+  
+  becomes:
+  
+  ```csharp
+  var descriptors = new AssemblyInspector()
+  	.FromAssemblyContaining<TestService>()
+  	.BasedOn(typeof(TestService))
+  	.WithServiceBase()
+  	.Configure(cfg => cfg.DependsOn = new Dependency[]
+  	{
+  		Parameter.ForKey("param").Eq("nonexisting")
+  	})
+  	.LifestyleTransient();
+  ```
+
 ## 0.1.2
 
 - Fixed namespaces.